+++ android/app/src/main/kotlin/com/example/builtin_inherittedwidget_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.builtin_inherittedwidget_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()