+++ lib/app/view/app_view.dart
@@ -6,10 +6,17 @@ import '/src/view.dart';
 /// App
 class ExampleApp extends AppStatefulWidget {
   ///
-  ExampleApp({super.key});
+  ExampleApp({super.key})
+      : super(
+          // splashScreen: const SplashScreen(),
+          inSplashScreen: () => const SplashScreen(),
+        );
   // This is the 'App State object' of the application.
   @override
   AppState createAppState() => _ExampleAppState();
+
+  @override
+  Widget? onSplashScreen(BuildContext context) => const SplashScreen();
 }
 
 /// This is the 'View' of the application.