+++ lib/app/view_app.dart
@@ -1,3 +1,8 @@
+//
 export 'view/menu/app_menu.dart';
 
+//
 export 'view/app_view.dart';
+
+//
+export 'view/splash_screen.dart';