+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.1.0
+August 21, 2024
+- Popup menu enable & disable built-in options
+- Supply a Splash Screen
+
 ## 1.0.1
 August 15, 2024
 - Corrected ImageAPIStateX and removed